+++ src/main/java/com/github/havlli/EventPilot/entity/embedtype/EmbedTypeRepository.java
@@ -2,5 +2,5 @@ package com.github.havlli.EventPilot.entity.embedtype;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface EmbedTypeRepository extends JpaRepository<EmbedType, Integer> {
+public interface EmbedTypeRepository extends JpaRepository<EmbedType, Long> {
 }
